requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64

ca-certificates=2019.1.23=0
certifi=2019.3.9=py37_0
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=8.2.0=hdf63c60_1
libstdcxx-ng=8.2.0=hdf63c60_1
maxminddb=1.4.1=pypi_0
maxminddb-geolite2=2018.703=pypi_0
ncurses=6.1=he6710b0_1
openssl=1.1.1b=h7b6447c_1
pip=19.0.3=py37_0
python=3.7.2=h0371630_0
python-geoip=1.2=pypi_0
python-geoip-geolite2=2015.303=pypi_0
readline=7.0=h7b6447c_5
setuptools=40.8.0=py37_0
sqlite=3.27.2=h7b6447c_0
tk=8.6.8=hbc83047_0
wheel=0.33.1=py37_0
xz=5.2.4=h14c3975_4
zlib=1.2.11=h7b6447c_3
